Restructure command files.
