+++ src/Commands/User/EchoCommand.php
@@ -8,22 +8,22 @@
  * file that was distributed with this source code.
  */
 
-namespace Longman\TelegramBot\Commands;
+namespace Longman\TelegramBot\Commands\UserCommands;
 
-use Longman\TelegramBot\Command;
+use Longman\TelegramBot\Commands\UserCommand;
 use Longman\TelegramBot\Request;
 
 /**
- * User "/slap" command
+ * User "/echo" command
  */
-class SlapCommand extends Command
+class EchoCommand extends UserCommand
 {
     /**#@+
      * {@inheritdoc}
      */
-    protected $name = 'slap';
-    protected $description = 'Slap someone with their username';
-    protected $usage = '/slap <@user>';
+    protected $name = 'echo';
+    protected $description = 'Show text';
+    protected $usage = '/echo <text>';
     protected $version = '1.0.1';
     /**#@-*/
 
@@ -35,20 +35,8 @@ class SlapCommand extends Command
     public function execute()
     {
         $message = $this->getMessage();
-
         $chat_id = $message->getChat()->getId();
-        $message_id = $message->getMessageId();
         $text = $message->getText(true);
-
-        $sender = '@' . $message->getFrom()->getUsername();
-
-        //username validation
-        $test = preg_match('/@[\w_]{5,}/', $text);
-        if ($test === 0) {
-            $text = $sender . ' sorry no one to slap around..';
-        } else {
-            $text = $sender . ' slaps ' . $text . ' around a bit with a large trout';
-        }
 
         $data = [
             'chat_id' => $chat_id,
